+++ src/constants/commonOptions.js
@@ -342,5 +342,20 @@ module.exports = {
     label: 'Use Delimiter',
     type: 'checkbox',
     tooltip: 'Separate thousands by local delimiter.'
+  },
+  'autofocus': {
+    label: 'Initial Focus',
+    type: 'checkbox',
+    tooltip: 'Make this field the initially focused element on this form.'
+  },
+  'collapsible': {
+    label: 'Collapsible',
+    type: 'checkbox',
+    tooltip: 'It\'s possible to hide body clicking the title.'
+  },
+  'collapsed': {
+    label: 'Collapsed',
+    type: 'checkbox',
+    tooltip: 'Component\'s body initially hidden.'
   }
 };